Compare amount in invoice currency against amount_to_pay
The transition_pay should have the same test as the default_ask because
amount_to_pay is computed in the invoice currency. Otherwise if the rate is
lower the test fail in transition_pay and the empty ask state is tested which
raises an exception.

issue7192
review52471002
(grafted from 2cbd0314a9c0b3fcc4b5bdf153e43eab8c3560ab)